Finally figured out how to add an autoconf option. Added the ability to disable AES-NI, and split cryptonight.c into four files - a header, functions common to both implementations, the AES-NI implementation, and the slow implementation. Also moved AES-NI specific asm to its own file.
