+++ configure.ac
@@ -1,7 +1,7 @@
-AC_INIT([cpuminer], [2.3.3])
+AC_INIT([cpuminer],[2.3.3])
 
-AC_PREREQ([2.59c])
-AC_CANONICAL_SYSTEM
+AC_PREREQ([2.69])
+AC_CANONICAL_TARGET
 AC_CONFIG_SRCDIR([cpu-miner.c])
 AM_INIT_AUTOMAKE([foreign])
 AC_CONFIG_HEADERS([cpuminer-config.h])
@@ -95,6 +95,12 @@ then
   )
 fi
 
+AC_ARG_ENABLE(aes-ni, AS_HELP_STRING([--disable-aes-ni], [disables aes-ni - use if your cpu is old]), use_hwaes=$enableval, use_hwaes=yes)
+
+if test "$use_hwaes" != "yes"; then
+  AC_DEFINE(USE_LOBOTOMIZED_AES, 1, [Set to one if we have to use the slow version])
+fi
+
 AC_CHECK_LIB(jansson, json_loads, request_jansson=false, request_jansson=true)
 AC_CHECK_LIB([pthread], [pthread_create], PTHREAD_LIBS="-lpthread",
   AC_CHECK_LIB([pthreadGC2], [pthread_create], PTHREAD_LIBS="-lpthreadGC2",
@@ -109,6 +115,7 @@ AM_CONDITIONAL([HAVE_WINDOWS], [test x$have_win32 = xtrue])
 AM_CONDITIONAL([ARCH_x86], [test x$have_x86 = xtrue])
 AM_CONDITIONAL([ARCH_x86_64], [test x$have_x86_64 = xtrue])
 AM_CONDITIONAL([ARCH_ARM], [test x$have_arm = xtrue])
+AM_CONDITIONAL([USE_LOBOTOMIZED_AES], test "$use_hwaes" = "no")
 
 if test x$request_jansson = xtrue
 then
