+++ cryptonight_common.c
@@ -1,0 +1,70 @@
+// Copyright (c) 2012-2013 The Cryptonote developers
+// Distributed under the MIT/X11 software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+// Modified for CPUminer by Lucas Jones
+
+#include "cpuminer-config.h"
+#include "miner.h"
+#include "crypto/c_keccak.h"
+#include "crypto/c_groestl.h"
+#include "crypto/c_blake256.h"
+#include "crypto/c_jh.h"
+#include "crypto/c_skein.h"
+#include "crypto/int-util.h"
+#include "crypto/hash-ops.h"
+#include "cryptonight.h"
+
+void do_blake_hash(const void* input, size_t len, char* output) {
+    blake256_hash((uint8_t*)output, input, len);
+}
+
+void do_groestl_hash(const void* input, size_t len, char* output) {
+    groestl(input, len * 8, (uint8_t*)output);
+}
+
+void do_jh_hash(const void* input, size_t len, char* output) {
+    jh_hash(HASH_SIZE * 8, input, 8 * len, (uint8_t*)output);
+}
+
+void do_skein_hash(const void* input, size_t len, char* output) {
+    skein_hash(8 * HASH_SIZE, input, 8 * len, (uint8_t*)output);
+}
+
+void xor_blocks_dst(const uint8_t* a, const uint8_t* b, uint8_t* dst) {
+    ((uint64_t*) dst)[0] = ((uint64_t*) a)[0] ^ ((uint64_t*) b)[0];
+    ((uint64_t*) dst)[1] = ((uint64_t*) a)[1] ^ ((uint64_t*) b)[1];
+}
+
+void (* const extra_hashes[4])(const void *, size_t, char *) = {do_blake_hash, do_groestl_hash, do_jh_hash, do_skein_hash};
+
+void cryptonight_hash(void* output, const void* input, size_t len) {
+    struct cryptonight_ctx *ctx = (struct cryptonight_ctx*)malloc(sizeof(struct cryptonight_ctx));
+    cryptonight_hash_ctx(output, input, len, ctx);
+    free(ctx);
+}
+
+int scanhash_cryptonight(int thr_id, uint32_t *pdata, const uint32_t *ptarget,
+        uint32_t max_nonce, unsigned long *hashes_done) {
+    uint32_t *nonceptr = (uint32_t*) (((char*)pdata) + 39);
+    uint32_t n = *nonceptr - 1;
+    const uint32_t first_nonce = n + 1;
+    const uint32_t Htarg = ptarget[7];
+    uint32_t hash[HASH_SIZE / 4] __attribute__((aligned(32)));
+
+    struct cryptonight_ctx *ctx = (struct cryptonight_ctx*)malloc(sizeof(struct cryptonight_ctx));
+
+	do {
+		*nonceptr = ++n;
+		cryptonight_hash_ctx(hash, pdata, 76, ctx);
+		if (unlikely(hash[7] < ptarget[7])) {
+			*hashes_done = n - first_nonce + 1;
+			free(ctx);
+			return true;
+		}
+	} while (likely((n <= max_nonce && !work_restart[thr_id].restart)));
+    
+    free(ctx);
+    *hashes_done = n - first_nonce + 1;
+    return 0;
+}
